requirements.txt
watchdog>=4.0.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-auth>=2.25.0
playwright>=1.40.0
mcp>=1.0.0
requests>=2.31.0